@@ -5,39 +5,39 @@ import React, { Component, PropTypes } from 'react';
 export default class Menu extends Component {
 
     static contextTypes = {
-        menuController: React.PropTypes.object
+        menuController: React.PropTypes.object,
     };
     static propTypes = {
-        children: PropTypes.node
+        children: PropTypes.node,
     };
 
     static Item = Item;
 
     constructor(props) {
         super(props);
-        this._id = null;
+        this.id = null;
     }
 
     componentDidMount() {
         setTimeout(() => {
-            this._id = this.context.menuController.add(this.refs.menu);
+            this.id = this.context.menuController.add(this.refs.menu);
         }, 0);
     }
     componentWillUnmount() {
         this.context.menuController.remove(this);
     }
     getId() {
-        return this._id;
+        return this.id;
     }
     getOptions(target) {
         const { children } = this.props;
 
         return (
             <Items
-              items={children}
-              closeMenu={this.close}
-              ref="menuItems"
-              target={target}
+                items={children}
+                closeMenu={this.close}
+                ref="menuItems"
+                target={target}
             />
         );
     }
