@@ -11,13 +11,13 @@ export default class Avatar extends Component {
         size: PropTypes.number,
         color: PropTypes.string,
         backgroundColor: PropTypes.string,
-        text: PropTypes.string
+        text: PropTypes.string,
     };
 
     static defaultProps = {
         size: 40,
         color: '#ffffff',
-        backgroundColor: getColor('paperGrey500')
+        backgroundColor: getColor('paperGrey500'),
     };
 
     render() {
@@ -30,8 +30,8 @@ export default class Avatar extends Component {
                     height: size,
                     borderRadius: size / 2,
                     borderColor: 'rgba(0,0,0,.1)',
-                    borderWidth: 1
-                }
+                    borderWidth: 1,
+                },
             });
         }
 
@@ -41,7 +41,7 @@ export default class Avatar extends Component {
             borderRadius: size / 2,
             backgroundColor: getColor(backgroundColor),
             alignItems: 'center',
-            justifyContent: 'center'
+            justifyContent: 'center',
         };
 
         if (icon) {
@@ -49,9 +49,9 @@ export default class Avatar extends Component {
                 <View style={{ flex: 1 }}>
                     <View style={style} >
                         <Icon
-                          name={icon}
-                          color={color}
-                          size={0.6 * size}
+                            name={icon}
+                            color={color}
+                            size={0.6 * size}
                         />
                     </View>
                 </View>
