@@ -11,12 +11,12 @@ const styles = {
         justifyContent: 'center',
         paddingVertical: 6,
         paddingHorizontal: 16,
-        borderRadius: 2
+        borderRadius: 2,
     },
     text: {
         position: 'relative',
-        top: 2
-    }
+        top: 2,
+    },
 };
 
 export default class Button extends Component {
@@ -28,26 +28,26 @@ export default class Button extends Component {
         overrides: PropTypes.shape({
             textColor: PropTypes.string,
             backgroundColor: PropTypes.string,
-            rippleColor: PropTypes.string
+            rippleColor: PropTypes.string,
         }),
         disabled: PropTypes.bool,
         raised: PropTypes.bool,
         onPress: PropTypes.func,
         onLongPress: PropTypes.func,
-        value: PropTypes.string
+        value: PropTypes.string,
     };
 
     static defaultProps = {
         theme: 'light',
         primary: PRIMARY,
         disabled: false,
-        raised: false
+        raised: false,
     };
 
     constructor(props) {
         super(props);
         this.state = {
-            elevation: 2
+            elevation: 2,
         };
     }
     onPress = () => {
@@ -59,13 +59,13 @@ export default class Button extends Component {
     }
     setElevation = () => {
         this.setState({
-            elevation: 4
+            elevation: 4,
         });
     };
 
     removeElevation = () => {
         this.setState({
-            elevation: 2
+            elevation: 2,
         });
     };
 
@@ -79,46 +79,46 @@ export default class Button extends Component {
             overrides,
             disabled,
             raised,
-            onLongPress
+            onLongPress,
         } = this.props;
 
         const textStyleMap = {
             flat: {
                 light: {
                     normal: {
-                        color: getColor(primary)
+                        color: getColor(primary),
                     },
                     disabled: {
-                        color: 'rgba(0,0,0,.26)'
-                    }
+                        color: 'rgba(0,0,0,.26)',
+                    },
                 },
                 dark: {
                     normal: {
-                        color: getColor(primary)
+                        color: getColor(primary),
                     },
                     disabled: {
-                        color: 'rgba(255,255,255,.3)'
-                    }
-                }
+                        color: 'rgba(255,255,255,.3)',
+                    },
+                },
             },
             raised: {
                 light: {
                     normal: {
-                        color: getColor(primary)
+                        color: getColor(primary),
                     },
                     disabled: {
-                        color: 'rgba(0,0,0,.26)'
-                    }
+                        color: 'rgba(0,0,0,.26)',
+                    },
                 },
                 dark: {
                     normal: {
-                        color: '#fff'
+                        color: '#fff',
                     },
                     disabled: {
-                        color: 'rgba(255,255,255,.3)'
-                    }
-                }
-            }
+                        color: 'rgba(255,255,255,.3)',
+                    },
+                },
+            },
         };
 
         const buttonStyleMap = {
@@ -131,45 +131,45 @@ export default class Button extends Component {
                     disabled: {
                         backgroundColor: 'rgba(0,0,0,.12)',
                         borderWidth: 1,
-                        borderColor: 'rgba(0,0,0,.12)'
-                    }
+                        borderColor: 'rgba(0,0,0,.12)',
+                    },
                 },
                 dark: {
                     normal: {
                         backgroundColor: getColor(primary),
                         borderWidth: 1,
-                        borderColor: 'rgba(0,0,0,.12)'
+                        borderColor: 'rgba(0,0,0,.12)',
                     },
                     disabled: {
                         backgroundColor: 'rgba(255,255,255,.12)',
                         borderWidth: 1,
-                        borderColor: 'rgba(0,0,0,.12)'
-                    }
-                }
-            }
+                        borderColor: 'rgba(0,0,0,.12)',
+                    },
+                },
+            },
         };
 
         const rippleColorMap = {
             flat: {
                 light: {
                     normal: 'rgba(153,153,153,.4)',
-                    disabled: 'rgba(0,0,0,0.06)'
+                    disabled: 'rgba(0,0,0,0.06)',
                 },
                 dark: {
                     normal: 'rgba(204,204,204,.25)',
-                    disabled: 'rgba(255,255,255,0.06)'
-                }
+                    disabled: 'rgba(255,255,255,0.06)',
+                },
             },
             raised: {
                 light: {
                     normal: 'rgba(153,153,153,.4)',
-                    disabled: 'rgba(0,0,0,.06)'
+                    disabled: 'rgba(0,0,0,.06)',
                 },
                 dark: {
                     normal: getColor(`${primary}700`),
-                    disabled: 'rgba(255,255,255,.06)'
-                }
-            }
+                    disabled: 'rgba(255,255,255,.06)',
+                },
+            },
         };
 
         const type = disabled ? 'disabled' : 'normal';
@@ -209,12 +209,12 @@ export default class Button extends Component {
         if (disabled) {
             return (
                 <View
-                  style={[
-                      styles.button,
-                      buttonStyle, {
-                          backgroundColor: buttonStyle && buttonStyle.backgroundColor
-                      }
-                  ]}
+                    style={[
+                        styles.button,
+                        buttonStyle, {
+                            backgroundColor: buttonStyle && buttonStyle.backgroundColor,
+                        },
+                    ]}
                 >
                     <Text style={[TYPO.paperFontButton, textStyle, styles.text]}>
                         {text || value}
@@ -226,19 +226,19 @@ export default class Button extends Component {
         if (!isCompatible('TouchableNativeFeedback')) {
             return (
                 <Ripple
-                  elevation={raised ? [2, 4] : null}
-                  rippleColor={rippleColor}
-                  onPress={!disabled ? this.onPress : null}
-                  onLongPress={!disabled ? onLongPress : null}
-                  style={[
-                      styles.button,
-                      buttonStyle, {
-                          backgroundColor: buttonStyle && buttonStyle.backgroundColor,
-                      }, raised && !isCompatible('elevation') && {
-                          borderWidth: 1,
-                          borderColor: 'rgba(0,0,0,.12)'
-                      }
-                  ]}
+                    elevation={raised ? [2, 4] : null}
+                    rippleColor={rippleColor}
+                    onPress={!disabled ? this.onPress : null}
+                    onLongPress={!disabled ? onLongPress : null}
+                    style={[
+                        styles.button,
+                        buttonStyle, {
+                            backgroundColor: buttonStyle && buttonStyle.backgroundColor,
+                        }, raised && !isCompatible('elevation') && {
+                            borderWidth: 1,
+                            borderColor: 'rgba(0,0,0,.12)',
+                        },
+                    ]}
                 >
                     <Text style={[TYPO.paperFontButton, textStyle]}>
                         {text || value}
@@ -249,18 +249,20 @@ export default class Button extends Component {
 
         return (
             <TouchableNativeFeedback
-              background={TouchableNativeFeedback.Ripple(rippleColor)}
-              onPress={!disabled ? this.onPress : null}
-              onLongPress={!disabled ? onLongPress : null}
-              onPressIn={raised ? this.setElevation : null}
-              onPressOut={raised ? this.removeElevation : null}
+                background={TouchableNativeFeedback.Ripple(rippleColor)}
+                onPress={!disabled ? this.onPress : null}
+                onLongPress={!disabled ? onLongPress : null}
+                onPressIn={raised ? this.setElevation : null}
+                onPressOut={raised ? this.removeElevation : null}
             >
-                <View style={[
-                    styles.button,
-                    buttonStyle, {
-                        backgroundColor: buttonStyle && buttonStyle.backgroundColor,
-                        elevation: raised ? elevation : 0
-                    }]}
+                <View
+                    style={[
+                        styles.button,
+                        buttonStyle, {
+                            backgroundColor: buttonStyle && buttonStyle.backgroundColor,
+                            elevation: raised ? elevation : 0,
+                        },
+                    ]}
                 >
                     <Text style={[TYPO.paperFontButton, textStyle, styles.text]}>
                         {text || value}
