@@ -5,22 +5,22 @@ const styles = StyleSheet.create({
     container: {
         flex: 1,
         paddingRight: -16,
-        paddingLeft: -16
+        paddingLeft: -16,
     },
     actions: {
-        flexDirection: 'row'
-    }
+        flexDirection: 'row',
+    },
 });
 
 export default class Actions extends Component {
 
     static propTypes = {
         position: PropTypes.oneOf(['left', 'right']),
-        children: PropTypes.node.isRequired
+        children: PropTypes.node.isRequired,
     };
 
     static defaultProps = {
-        position: 'left'
+        position: 'left',
     };
 
     render() {
@@ -29,10 +29,10 @@ export default class Actions extends Component {
         return (
             <View style={styles.container}>
                 <View
-                  style={[
-                      styles.actions,
-                      { alignSelf: position === 'left' ? 'flex-start' : 'flex-end' }
-                  ]}
+                    style={[
+                        styles.actions,
+                      { alignSelf: position === 'left' ? 'flex-start' : 'flex-end' },
+                    ]}
                 >
                     {children}
                 </View>
