@@ -15,8 +15,8 @@ const styles = StyleSheet.create({
         paddingTop: 24,
         paddingBottom: 16,
         paddingLeft: 16,
-        paddingRight: 16
-    }
+        paddingRight: 16,
+    },
 });
 
 export default class Media extends Component {
@@ -25,12 +25,12 @@ export default class Media extends Component {
         image: PropTypes.shape({ type: PropTypes.oneOf([Image]) }).isRequired,
         height: PropTypes.number,
         overlay: PropTypes.bool,
-        children: PropTypes.node
+        children: PropTypes.node,
     };
 
     static defaultProps = {
         height: 150,
-        overlay: false
+        overlay: false,
     };
 
     render() {
@@ -39,14 +39,14 @@ export default class Media extends Component {
         return (
             <View style={{ height }}>
                 {React.cloneElement(image, {
-                    style: [styles.media, { height }]
+                    style: [styles.media, { height }],
                 })}
                 {children &&
                     <View
-                      style={[
-                          styles.content,
-                          overlay && { backgroundColor: 'rgba(0,0,0,.35)' }
-                      ]}
+                        style={[
+                            styles.content,
+                            overlay && { backgroundColor: 'rgba(0,0,0,.35)' },
+                        ]}
                     >
                         {children}
                     </View>
