@@ -15,8 +15,8 @@ const styles = {
         marginVertical: 8,
         marginHorizontal: 8,
         paddingLeft: 16,
-        paddingRight: 16
-    }
+        paddingRight: 16,
+    },
 };
 
 export default class Card extends Component {
@@ -25,19 +25,19 @@ export default class Card extends Component {
         theme: PropTypes.string,
         overrides: PropTypes.shape({
             backgroundColor: PropTypes.string,
-            rippleColor: PropTypes.string
+            rippleColor: PropTypes.string,
         }),
         elevation: PropTypes.number,
         fullWidth: PropTypes.bool,
         disabled: PropTypes.bool,
         onPress: PropTypes.func,
         children: PropTypes.node.isRequired,
-        style: PropTypes.object
+        style: PropTypes.object,
     };
 
     static defaultProps = {
         elevation: 2,
-        disabled: false
+        disabled: false,
     };
 
     static Media = Media;
@@ -55,22 +55,22 @@ export default class Card extends Component {
             disabled,
             onPress,
             children,
-            style
+            style,
         } = this.props;
 
         const cardStyle = (() => ([
             styles.container, {
-                elevation
+                elevation,
             }, fullWidth && {
-                marginHorizontal: 0
+                marginHorizontal: 0,
             }, !isCompatible('elevation') && {
                 borderWidth: 1,
-                borderColor: 'rgba(0,0,0,.12)'
+                borderColor: 'rgba(0,0,0,.12)',
             }, theme && {
-                backgroundColor: COLOR[theme].color
+                backgroundColor: COLOR[theme].color,
             }, overrides && overrides.backgroundColor && {
-                backgroundColor: overrides.backgroundColor
-            }, style
+                backgroundColor: overrides.backgroundColor,
+            }, style,
         ]))();
 
 
@@ -94,8 +94,8 @@ export default class Card extends Component {
         if (!isCompatible('TouchableNativeFeedback')) {
             return (
                 <Ripple
-                  rippleColor={rippleColor}
-                  onPress={onPress}
+                    rippleColor={rippleColor}
+                    onPress={onPress}
                 >
                     <View style={cardStyle}>
                         {children}
@@ -106,8 +106,8 @@ export default class Card extends Component {
 
         return (
             <TouchableNativeFeedback
-              background={TouchableNativeFeedback.Ripple(rippleColor)}
-              onPress={onPress}
+                background={TouchableNativeFeedback.Ripple(rippleColor)}
+                onPress={onPress}
             >
                 <View style={cardStyle}>
                     {children}
