@@ -10,13 +10,13 @@ const styles = StyleSheet.create({
     container: {
         flexDirection: 'row',
         alignItems: 'center',
-        backgroundColor: 'rgba(0,0,0,0)'
+        backgroundColor: 'rgba(0,0,0,0)',
     },
     label: {
         marginLeft: 16,
         opacity: COLOR.darkPrimaryOpacity.opacity,
-        flex: 1
-    }
+        flex: 1,
+    },
 });
 
 export default class Checkbox extends Component {
@@ -28,16 +28,16 @@ export default class Checkbox extends Component {
         value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
         checked: PropTypes.bool,
         disabled: PropTypes.bool,
-        onCheck: PropTypes.func
+        onCheck: PropTypes.func,
     };
 
     static defaultProps = {
         theme: 'light',
         primary: PRIMARY,
-        disabled: false
+        disabled: false,
     };
 
-    _onPress = () => {
+    onPress = () => {
         const { checked, disabled, onCheck, value } = this.props;
         if (disabled === false) {
             onCheck(!checked, value);
@@ -60,36 +60,36 @@ export default class Checkbox extends Component {
             light: {
                 disabled: '#000000',
                 checked: COLOR[`${primary}500`].color,
-                default: '#000000'
+                default: '#000000',
             },
             dark: {
                 disabled: '#ffffff',
                 checked: COLOR[`${primary}500`].color,
-                default: '#ffffff'
-            }
+                default: '#ffffff',
+            },
         };
 
         const opacityMap = {
             light: {
                 checked: 1,
                 default: 0.54,
-                disabled: 0.26
+                disabled: 0.26,
             },
             dark: {
                 checked: 1,
                 default: 0.7,
-                disabled: 0.3
-            }
+                disabled: 0.3,
+            },
         };
 
         const underlayMap = {
             light: 'rgba(0,0,0,.12)',
-            dark: 'rgba(255,255,255,.12)'
+            dark: 'rgba(255,255,255,.12)',
         };
 
         const labelColorMap = {
             light: '#000000',
-            dark: '#ffffff'
+            dark: '#ffffff',
         };
 
         const CURR_COLOR = colorMap[theme][status];
@@ -99,39 +99,40 @@ export default class Checkbox extends Component {
 
         return (
             <TouchableHighlight
-              onPress={this._onPress}
-              underlayColor={disabled ? 'rgba(0,0,0,0)' : UNDERLAY_COLOR}
-              activeOpacity={1}
+                onPress={this.onPress}
+                underlayColor={disabled ? 'rgba(0,0,0,0)' : UNDERLAY_COLOR}
+                activeOpacity={1}
             >
                 <View style={styles.container}>
                     <IconToggle
-                      disabled={disabled}
-                      color={CURR_COLOR}
-                      onPress={this._onPress}
+                        disabled={disabled}
+                        color={CURR_COLOR}
+                        onPress={this.onPress}
                     >
-                        <Icon name={checked ? 'check-box' : 'check-box-outline-blank'}
-                          size={24}
-                          color={CURR_COLOR}
-                          key={value}
-                          style={{
-                              opacity: OPACITY,
-                              margin: 16,
-                          }}
+                        <Icon
+                            name={checked ? 'check-box' : 'check-box-outline-blank'}
+                            size={24}
+                            color={CURR_COLOR}
+                            key={value}
+                            style={{
+                                opacity: OPACITY,
+                                margin: 16,
+                            }}
                         />
                     </IconToggle>
                     <View
-                      style={styles.labelContainer}
-                      onPress={this._onPress}
+                        style={styles.labelContainer}
+                        onPress={this.onPress}
                     >
                         <Text
-                          style={[
-                              typos.paperFontBody1,
-                              styles.label,
-                              COLOR[`${theme}PrimaryOpacity`],
-                              disabled && COLOR[`${theme}DisabledOpacity`], {
-                                  color: LABEL_COLOR
-                              }
-                          ]}
+                            style={[
+                                typos.paperFontBody1,
+                                styles.label,
+                                COLOR[`${theme}PrimaryOpacity`],
+                                disabled && COLOR[`${theme}DisabledOpacity`], {
+                                    color: LABEL_COLOR,
+                                },
+                            ]}
                         >
                             {this.props.label}
                         </Text>
