@@ -13,19 +13,19 @@ export default class CheckboxGroup extends Component {
         items: PropTypes.arrayOf(PropTypes.shape({
             value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
             label: PropTypes.string,
-            disabled: PropTypes.bool
-        }))
+            disabled: PropTypes.bool,
+        })),
     };
 
     static defaultProps = {
         theme: 'light',
-        primary: PRIMARY
+        primary: PRIMARY,
     };
 
     constructor(props) {
         super(props);
         this.state = {
-            selected: []
+            selected: [],
         };
     }
 
@@ -37,20 +37,20 @@ export default class CheckboxGroup extends Component {
         }
     };
 
-    _onChange = (checked, value) => {
+    onChange = (checked, value) => {
         const { selected } = this.state;
 
         if (checked) {
             this.setState({
-                selected: [...selected, value]
+                selected: [...selected, value],
             });
         } else {
             const index = selected.indexOf(value);
             this.setState({
                 selected: [
                     ...selected.slice(0, index),
-                    ...selected.slice(index + 1)
-                ]
+                    ...selected.slice(index + 1),
+                ],
             });
         }
 
@@ -75,7 +75,7 @@ export default class CheckboxGroup extends Component {
      */
     set value(value) {
         this.setState({
-            selected: value
+            selected: value,
         });
 
         const { onSelect } = this.props;
@@ -96,14 +96,14 @@ export default class CheckboxGroup extends Component {
                         const { value } = item;
                         return (
                             <Checkbox
-                              ref={value}
-                              key={`Checkbox${value}`}
-                              value={value}
-                              theme={theme}
-                              primary={primary}
-                              onCheck={this._onChange}
-                              checked={selected && selected.indexOf(value) !== -1}
-                              {...item}
+                                ref={value}
+                                key={`Checkbox${value}`}
+                                value={value}
+                                theme={theme}
+                                primary={primary}
+                                onCheck={this.onChange}
+                                checked={selected && selected.indexOf(value) !== -1}
+                                {...item}
                             />
                         );
                     })
