@@ -13,7 +13,7 @@ const defaultStyles = {
     },
     innerContainer: {
         margin: 8,
-    }
+    },
 };
 
 class DialogFooter extends Component {
