@@ -9,7 +9,7 @@ const defaultStyles = {
     dialogContainer: {
         backgroundColor: '#ffffff',
         paddingBottom: 24,
-    }
+    },
 };
 
 class DialogContent extends Component {
