@@ -15,7 +15,7 @@ const defaultStyles = {
         width: 280,
         paddingTop: 24,
         paddingHorizontal: 24,
-    }
+    },
 };
 
 export default class Dialog extends Component {
@@ -24,14 +24,14 @@ export default class Dialog extends Component {
         theme: PropTypes.string,
         overrides: PropTypes.shape({
             backgroundColor: PropTypes.string,
-            rippleColor: PropTypes.string
+            rippleColor: PropTypes.string,
         }),
         elevation: PropTypes.number,
         fullWidth: PropTypes.bool,
         disabled: PropTypes.bool,
         onPress: PropTypes.func,
         children: PropTypes.node.isRequired,
-        style: PropTypes.object
+        style: PropTypes.object,
     };
 
     static Title = Title;
