@@ -13,7 +13,7 @@ const defaultStyles = {
     },
     actionContainer: {
         marginLeft: 8,
-    }
+    },
 };
 
 class DialogDefaultActions extends Component {
