@@ -14,7 +14,7 @@ const defaultStyles = {
         fontSize: 20,
         fontWeight: 'bold',
         color: 'black',
-    }
+    },
 };
 
 class DialogHeader extends Component {
