@@ -4,8 +4,8 @@ import React, { Component, PropTypes } from 'react';
 
 const styles = {
     divider: {
-        height: 1
-    }
+        height: 1,
+    },
 };
 
 export default class Divider extends Component {
@@ -13,12 +13,12 @@ export default class Divider extends Component {
     static propTypes = {
         inset: PropTypes.bool,
         theme: PropTypes.oneOf(THEME_NAME),
-        style: PropTypes.object
+        style: PropTypes.object,
     };
 
     static defaultProps = {
         inset: false,
-        theme: 'light'
+        theme: 'light',
     };
 
     render() {
@@ -28,7 +28,7 @@ export default class Divider extends Component {
 
         return (
             <View
-              style={[styles.divider, inset && { marginLeft: 72 }, { backgroundColor }, style]}
+                style={[styles.divider, inset && { marginLeft: 72 }, { backgroundColor }, style]}
             />
         );
     }
