@@ -8,8 +8,8 @@ const styles = {
         position: 'absolute',
         top: 0,
         left: 0,
-        right: 0
-    }
+        right: 0,
+    },
 };
 
 export default class Header extends Component {
@@ -20,12 +20,12 @@ export default class Header extends Component {
         image: PropTypes.shape({ type: PropTypes.oneOf([Image]) }),
         backgroundColor: PropTypes.string,
         height: PropTypes.number,
-        children: PropTypes.node
+        children: PropTypes.node,
     };
 
     static defaultProps = {
         height: 150,
-        backgroundColor: '#333333'
+        backgroundColor: '#333333',
     };
 
     render() {
@@ -33,7 +33,7 @@ export default class Header extends Component {
 
         const contetnStyle = {
             height,
-            backgroundColor: image ? 'transparent' : backgroundColor
+            backgroundColor: image ? 'transparent' : backgroundColor,
         };
 
         const content = (
