@@ -7,17 +7,17 @@ const styles = {
     accountContainer: {
         flex: 1,
         paddingHorizontal: 16,
-        marginBottom: 8
+        marginBottom: 8,
     },
     activeAvatarContainer: {
-        flex: 1
+        flex: 1,
     },
     inactiveAvatarContainer: {
         paddingLeft: 8,
     },
     topContainer: {
         flex: 1,
-        justifyContent: 'center'
+        justifyContent: 'center',
     },
     avatarsContainer: {
         flexDirection: 'row',
@@ -26,17 +26,17 @@ const styles = {
         height: 56,
         flexDirection: 'row',
         alignItems: 'center',
-        justifyContent: 'center'
+        justifyContent: 'center',
     },
     textContainer: {
-        flex: 1
+        flex: 1,
     },
     firstLineContainer: {
     },
     primaryText: {
         fontSize: 14,
-        fontWeight: 'bold'
-    }
+        fontWeight: 'bold',
+    },
 };
 
 export default class HeaderAcount extends Component {
@@ -44,7 +44,7 @@ export default class HeaderAcount extends Component {
     static propTypes = {
         accounts: PropTypes.arrayOf(PropTypes.shape({
             avatar: PropTypes.element,
-            onPress: PropTypes.func
+            onPress: PropTypes.func,
         })),
         avatar: PropTypes.element,
         height: PropTypes.number,
@@ -52,10 +52,10 @@ export default class HeaderAcount extends Component {
         onRightIconPress: PropTypes.func,
         primaryText: PropTypes.string,
         rightIcon: PropTypes.element,
-        secondaryText: PropTypes.string
+        secondaryText: PropTypes.string,
     };
     static defaultProps = {
-        height: 150
+        height: 150,
     };
 
     render() {
@@ -66,7 +66,7 @@ export default class HeaderAcount extends Component {
             secondaryText,
             rightIcon,
             onPress,
-            onRightIconPress
+            onRightIconPress,
         } = this.props;
 
         const primaryColor = 'rgba(0,0,0,.87)';
@@ -100,9 +100,9 @@ export default class HeaderAcount extends Component {
                             </Text>
                             {
                                 secondaryText &&
-                                <Text>
-                                    {secondaryText}
-                                </Text>
+                                    <Text>
+                                        {secondaryText}
+                                    </Text>
                             }
                         </View>
                         <View style={[styles.iconContainer]}>
