@@ -8,31 +8,31 @@ import Ripple from '../polyfill/Ripple';
 const styles = {
     section: {
         flex: 1,
-        marginTop: 8
+        marginTop: 8,
     },
     item: {
         flex: 1,
         flexDirection: 'row',
         alignItems: 'center',
         height: 48,
-        paddingLeft: 16
+        paddingLeft: 16,
     },
     subheader: {
         flex: 1,
     },
     icon: {
         position: 'absolute',
-        top: 13
+        top: 13,
     },
     value: {
         flex: 1,
         paddingLeft: 56,
-        top: 2
+        top: 2,
     },
     label: {
         paddingRight: 16,
-        top: 2
-    }
+        top: 2,
+    },
 };
 
 export default class Section extends Component {
@@ -47,18 +47,18 @@ export default class Section extends Component {
             onPress: PropTypes.func,
             onLongPress: PropTypes.func,
             active: PropTypes.bool,
-            disabled: PropTypes.bool
-        }))
+            disabled: PropTypes.bool,
+        })),
     };
 
     renderRow = (item, index, color) => (
         <View key={index} style={styles.item} >
             {item.icon &&
                 <Icon
-                  name={item.icon}
-                  color={color}
-                  size={22}
-                  style={styles.icon}
+                    name={item.icon}
+                    color={color}
+                    size={22}
+                    style={styles.icon}
                 />
             }
             <View style={styles.value}>
@@ -82,12 +82,12 @@ export default class Section extends Component {
         const textStyleMap = {
             light: {
                 default: 'rgba(0,0,0,.87)',
-                disabled: 'rgba(0,0,0,.38)'
+                disabled: 'rgba(0,0,0,.38)',
             },
             dark: {
                 default: '#ffffff',
-                disabled: 'rgba(255,255,255,.30)'
-            }
+                disabled: 'rgba(255,255,255,.30)',
+            },
         };
 
         const subheaderStyleMap = {
@@ -121,15 +121,15 @@ export default class Section extends Component {
                     if (!isCompatible('TouchableNativeFeedback')) {
                         return (
                             <Ripple
-                              key={i}
-                              rippleColor="rgba(153,153,153,.4)"
-                              onPress={item.onPress}
-                              onLongPress={item.onLongPress}
-                              style={{
-                                  flex: 1,
-                                  flexDirection: 'row',
-                                  backgroundColor: item.active ? ACTIVE_COLOR : null
-                              }}
+                                key={i}
+                                rippleColor="rgba(153,153,153,.4)"
+                                onPress={item.onPress}
+                                onLongPress={item.onLongPress}
+                                style={{
+                                    flex: 1,
+                                    flexDirection: 'row',
+                                    backgroundColor: item.active ? ACTIVE_COLOR : null,
+                                }}
                             >
                                 {this.renderRow(item, i, TEXT_COLOR)}
                             </Ripple>
@@ -138,10 +138,10 @@ export default class Section extends Component {
 
                     return (
                         <TouchableNativeFeedback
-                          key={i}
-                          background={TouchableNativeFeedback.Ripple('rgba(153,153,153,.4)')}
-                          onPress={item.onPress}
-                          onLongPress={item.onLongPress}
+                            key={i}
+                            background={TouchableNativeFeedback.Ripple('rgba(153,153,153,.4)')}
+                            onPress={item.onPress}
+                            onLongPress={item.onLongPress}
                         >
                             <View style={item.active ? { backgroundColor: ACTIVE_COLOR } : {}}>
                                 {this.renderRow(item, i, TEXT_COLOR)}
