@@ -7,8 +7,8 @@ import Section from './Section';
 
 const styles = {
     container: {
-        flex: 1
-    }
+        flex: 1,
+    },
 };
 
 export default class Drawer extends Component {
@@ -17,14 +17,14 @@ export default class Drawer extends Component {
         theme: PropTypes.oneOf(THEME_NAME),
         primary: PropTypes.oneOf(PRIMARY_COLORS),
         overrides: PropTypes.shape({
-            backgroundColor: PropTypes.string
+            backgroundColor: PropTypes.string,
         }),
-        children: PropTypes.node.isRequired
+        children: PropTypes.node.isRequired,
     };
 
     static defaultProps = {
         theme: 'light',
-        primary: 'paperGrey'
+        primary: 'paperGrey',
     };
 
     static Header = Header;
@@ -36,7 +36,7 @@ export default class Drawer extends Component {
 
         const backgroundColorMap = {
             light: '#ffffff',
-            dark: '#333333'
+            dark: '#333333',
         };
 
         const backgroundColor = (() => {
