@@ -8,12 +8,12 @@ export default class Icon extends Component {
         name: PropTypes.string.isRequired,
         style: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
         size: PropTypes.number,
-        color: PropTypes.string
+        color: PropTypes.string,
     };
 
     static defaultProps = {
         size: 30,
-        color: '#757575'
+        color: '#757575',
     };
 
     render() {
@@ -21,10 +21,10 @@ export default class Icon extends Component {
 
         return (
             <VectorIcon
-              name={name}
-              size={size}
-              color={getColor(color)}
-              style={style}
+                name={name}
+                size={size}
+                color={getColor(color)}
+                style={style}
             />
         );
     }
