@@ -7,11 +7,11 @@ const styles = {
         position: 'absolute',
         borderRadius: 7.5,
         width: 15,
-        height: 15
+        height: 15,
     },
     badgeText: {
-        textAlign: 'center'
-    }
+        textAlign: 'center',
+    },
 };
 
 export default class IconToggle extends Component {
@@ -27,14 +27,14 @@ export default class IconToggle extends Component {
             value: PropTypes.number,
             animate: PropTypes.bool,
             backgroundColor: PropTypes.string,
-            textColor: PropTypes.string
-        })
+            textColor: PropTypes.string,
+        }),
     };
 
     static defaultProps = {
         opacity: .1,
         disabled: false,
-        percent: 90
+        percent: 90,
     };
 
     constructor(props) {
@@ -43,20 +43,20 @@ export default class IconToggle extends Component {
             scaleValue: new Animated.Value(0.001),
             opacityValue: new Animated.Value(0.1),
             maxOpacity: props.opacity,
-            size: null
+            size: null,
         };
-        this._responder = {
+        this.responder = {
             onStartShouldSetResponder: () => true,
-            onResponderGrant: this._highlight,
-            onResponderRelease: this._handleResponderEnd,
-            onResponderTerminate: this._unHighlight
+            onResponderGrant: this.highlight,
+            onResponderRelease: this.handleResponderEnd,
+            onResponderTerminate: this.unHighlight,
         };
     }
 
     componentDidUpdate(prevProps) {
         if (this.props && this.props.badge) {
             if (this.props.badge.value !== prevProps.badge.value) {
-                this._animateBadgeCounter();
+                this.animateBadgeCounter();
             }
         }
     }
@@ -65,7 +65,7 @@ export default class IconToggle extends Component {
         const { width, height } = event.nativeEvent.layout;
 
         this.setState({
-            size: width > height ? width : height
+            size: width > height ? width : height,
         });
     };
 
@@ -74,15 +74,15 @@ export default class IconToggle extends Component {
      *
      * @private
      */
-    _highlight = () => {
+    highlight = () => {
         if (!this.props.disabled) {
             Animated.timing(this.state.scaleValue, {
                 toValue: 1,
-                duration: 150
+                duration: 150,
             }).start();
             Animated.timing(this.state.opacityValue, {
                 toValue: this.state.maxOpacity,
-                duration: 100
+                duration: 100,
             }).start();
         }
     };
@@ -92,14 +92,14 @@ export default class IconToggle extends Component {
      *
      * @private
      */
-    _unHighlight = () => {
+    unHighlight = () => {
         if (!this.props.disabled) {
             Animated.timing(this.state.scaleValue, {
                 toValue: 0.001,
-                duration: 1500
+                duration: 1500,
             }).start();
             Animated.timing(this.state.opacityValue, {
-                toValue: 0
+                toValue: 0,
             }).start();
         }
     };
@@ -108,14 +108,14 @@ export default class IconToggle extends Component {
      *
      * @private
      */
-    _animateBadgeCounter = () => {
+    animateBadgeCounter = () => {
         if (this.badgeAnim && this.props.badge && this.props.badge.animate !== false) {
             Animated.spring(this.badgeAnim, {
                 toValue: 0,   // Returns to the start
                 velocity: 8,  // Velocity makes it move
                 tension: -5, // Slow
                 friction: 1,  // Oscillate a lot
-                duration: 300
+                duration: 300,
             }).start();
         }
     };
@@ -124,11 +124,11 @@ export default class IconToggle extends Component {
      *
      * @private
      */
-    _handleResponderEnd = () => {
+    handleResponderEnd = () => {
         const { disabled, onPress } = this.props;
 
         if (!disabled) {
-            this._unHighlight();
+            this.unHighlight();
 
             if (onPress) {
                 onPress();
@@ -167,60 +167,62 @@ export default class IconToggle extends Component {
         this.badgeAnim = this.badgeAnim || new Animated.Value(0);
 
         return (
-            <View {...this._responder}>
+            <View {...this.responder}>
                 <View>
                     {size &&
                         <Animated.View
-                          style={[{
-                              position: 'absolute',
-                              left: ((1 - percent) * size) / 2,
-                              top: ((1 - percent) * size) / 2,
-                              width: percent * size,
-                              height: percent * size,
-                              borderRadius: percent * size / 2,
-                              transform: [{ scale: scaleValue }],
-                              opacity: opacityValue,
-                              backgroundColor: getColor(color)
-                          }]}
+                            style={[{
+                                position: 'absolute',
+                                left: ((1 - percent) * size) / 2,
+                                top: ((1 - percent) * size) / 2,
+                                width: percent * size,
+                                height: percent * size,
+                                borderRadius: percent * size / 2,
+                                transform: [{ scale: scaleValue }],
+                                opacity: opacityValue,
+                                backgroundColor: getColor(color),
+                            }]}
                         />
                     }
                     <View style={{ backgroundColor: 'transparent' }} onLayout={this.setSize}>
                         {children}
                     </View>
                     {size && badge && typeof badge.value === 'number' &&
-                        <Animated.View style={[
-                            styles.badgeContainer, {
-                                backgroundColor: getColor(badge.backgroundColor),
-                                top: size / 10,
-                                right: size / 10,
-                                transform: [   // Array order matters
-                                { scale: this.badgeAnim.interpolate({
-                                    inputRange: [0, 1],
-                                    outputRange: [1, 1.25]
-                                }) },
-                                { translateX: this.badgeAnim.interpolate({
-                                    inputRange: [0, 1],
-                                    outputRange: [0, -6]
-                                }) },
-                                { translateY: this.badgeAnim.interpolate({
-                                    inputRange: [0, 1],
-                                    outputRange: [0, 5]
-                                }) },
-                                { rotate: this.badgeAnim.interpolate({
-                                    inputRange: [0, 1],
-                                    outputRange: [
-                                        '0deg', '45deg' // 'deg' or 'rad'
-                                    ]
-                                }) }
-                                ] }]}
+                        <Animated.View
+                            style={[
+                                styles.badgeContainer, {
+                                    backgroundColor: getColor(badge.backgroundColor),
+                                    top: size / 10,
+                                    right: size / 10,
+                                    transform: [   // Array order matters
+                                        { scale: this.badgeAnim.interpolate({
+                                            inputRange: [0, 1],
+                                            outputRange: [1, 1.25],
+                                        }) },
+                                        { translateX: this.badgeAnim.interpolate({
+                                            inputRange: [0, 1],
+                                            outputRange: [0, -6],
+                                        }) },
+                                        { translateY: this.badgeAnim.interpolate({
+                                            inputRange: [0, 1],
+                                            outputRange: [0, 5],
+                                        }) },
+                                        { rotate: this.badgeAnim.interpolate({
+                                            inputRange: [0, 1],
+                                            outputRange: [
+                                                '0deg', '45deg', // 'deg' or 'rad'
+                                            ],
+                                        }) },
+                                    ] },
+                            ]}
                         >
                             <View style={{ flex: 1, justifyContent: 'center' }}>
                                 <Text
-                                  style={[
-                                      styles.badgeText,
+                                    style={[
+                                        styles.badgeText,
                                       { color: getColor(badge.textColor) },
-                                      badge.value > 99 ? { fontSize: 8 } : { fontSize: 10 }
-                                  ]}
+                                        badge.value > 99 ? { fontSize: 8 } : { fontSize: 10 },
+                                    ]}
                                 >
                                     {badge.value}
                                 </Text>
