@@ -5,7 +5,7 @@ import {
     View,
     Text,
     TouchableWithoutFeedback,
-    TouchableNativeFeedback
+    TouchableNativeFeedback,
 } from 'react-native';
 import { TYPO } from './config';
 import Icon from './Icon';
@@ -18,7 +18,7 @@ const defaultStyles = StyleSheet.create({
         flex: 1,
         flexDirection: 'row',
         paddingLeft: 16,
-        paddingRight: 16
+        paddingRight: 16,
     },
 
     contentViewContainer: {
@@ -29,7 +29,7 @@ const defaultStyles = StyleSheet.create({
 
     leftViewContainer: {
         flexDirection: 'column',
-        width: 56
+        width: 56,
     },
     leftIcon: {
     },
@@ -37,43 +37,43 @@ const defaultStyles = StyleSheet.create({
     },
 
     textViewContainer: {
-        flex: 1
+        flex: 1,
     },
     primaryText: Object.assign({}, TYPO.paperFontSubhead, { lineHeight: 24 }),
     secondaryText: Object.assign({}, TYPO.paperFontBody1, {
         lineHeight: 22,
         fontSize: 14,
-        color: 'rgba(0,0,0,.54)'
+        color: 'rgba(0,0,0,.54)',
     }),
     firstLine: {
-        flexDirection: 'row'
+        flexDirection: 'row',
     },
     primaryTextContainer: {
-        flex: 1
+        flex: 1,
     },
 
     rightViewContainer: {
         flexDirection: 'column',
-        paddingLeft: 16
+        paddingLeft: 16,
     },
     rightActionsViewContainer: {
-        flex: 1
+        flex: 1,
     },
     rightIcon: {
         flex: 1,
     },
     rightAvatar: {
-        width: 56
+        width: 56,
     },
 
     captionText: Object.assign({}, TYPO.paperFontCaption),
     captionTextContainer: {
-        flex: 1
+        flex: 1,
     },
     captionTextContainer2: {
         alignSelf: 'flex-end',
-        alignItems: 'flex-end'
-    }
+        alignItems: 'flex-end',
+    },
 });
 
 export default class List extends Component {
@@ -85,7 +85,7 @@ export default class List extends Component {
         captionText: PropTypes.string,
         secondaryTextMoreLine: PropTypes.arrayOf(PropTypes.shape({
             text: PropTypes.string.isRequired,
-            style: PropTypes.object
+            style: PropTypes.object,
         })),
         leftIcon: PropTypes.element,
         onLeftIconClicked: PropTypes.func,
@@ -93,7 +93,7 @@ export default class List extends Component {
         leftAvatar: PropTypes.element,
         menuActions: PropTypes.shape({
             onPress: PropTypes.func,
-            labels: PropTypes.array
+            labels: PropTypes.array,
         }),
         rightAvatar: PropTypes.element,
         lines: React.PropTypes.oneOf([1, 2, 3, 'dynamic']),
@@ -101,32 +101,31 @@ export default class List extends Component {
         onPress: PropTypes.func,
         onPressValue: PropTypes.any,
         onLeftIconClick: PropTypes.func,
-        onRightIconClick: PropTypes.func
+        onRightIconClick: PropTypes.func,
     };
 
     static defaultProps = {
         lines: 1,
-        primaryColor: 'rgba(0,0,0,.87)'
+        primaryColor: 'rgba(0,0,0,.87)',
     };
-    _handleShowPopupError() { }
-    _onMenuPressed = () => {
+    onMenuPressed = () => {
         const { menuActions } = this.props;
 
         UIManager.showPopupMenu(
             findNodeHandle(this.refs.menu),
             menuActions.labels,
-            this._handleShowPopupError,
+            () => {},
             menuActions.onPress
         );
     };
-    _onListItemPressed = () => {
+    onListItemPressed = () => {
         const { onPress, onPressValue } = this.props;
 
         if (onPress) {
             onPress(onPressValue);
         }
     };
-    _onLeftIconPressed = () => {
+    onLeftIconPressed = () => {
         const { onLeftIconClicked, onPressValue } = this.props;
 
         if (onLeftIconClicked) {
@@ -134,11 +133,11 @@ export default class List extends Component {
         }
     };
 
-    _getStyles() {
+    getStyles() {
         const {
             dense,
             lines,
-            secondaryText
+            secondaryText,
         } = this.props;
 
         let numberOfLines = lines;
@@ -152,27 +151,27 @@ export default class List extends Component {
         if (numberOfLines === 3) {
             styles.listContainer = {
                 alignItems: 'center',
-                height: dense === true ? 76 : 88
+                height: dense === true ? 76 : 88,
             };
         } else if (numberOfLines === 2) {
             styles.listContainer = {
-                height: dense === true ? 60 : 72
+                height: dense === true ? 60 : 72,
             };
             styles.contentViewContainer = {
-                alignItems: 'center'
+                alignItems: 'center',
             };
         } else if (numberOfLines === 1) {
             styles.listContainer = {
-                height: dense === true ? 48 : 56
+                height: dense === true ? 48 : 56,
             };
             styles.contentViewContainer = {
-                alignItems: 'center'
+                alignItems: 'center',
             };
         } else if (numberOfLines === 'dynamic') {
             styles.listContainer = {
                 alignItems: 'center',
                 paddingTop: 16,
-                paddingBottom: 16
+                paddingBottom: 16,
             };
         }
 
@@ -194,128 +193,131 @@ export default class List extends Component {
             onRightIconClick,
             primaryColor,
             secondaryTextMoreLine,
-            captionText
+            captionText,
         } = this.props;
 
-        const styles = this._getStyles();
+        const styles = this.getStyles();
 
         const content = (
-                <View style={[defaultStyles.listContainer, styles.listContainer]}>
+            <View style={[defaultStyles.listContainer, styles.listContainer]}>
 
-                    <View style={[defaultStyles.contentViewContainer, styles.contentViewContainer]}>
+                <View style={[defaultStyles.contentViewContainer, styles.contentViewContainer]}>
 
-                        {
-                            (leftIcon || leftAvatar) &&
+                    {
+                        (leftIcon || leftAvatar) &&
                             <View
-                              style={[defaultStyles.leftViewContainer, styles.leftViewContainer]}
+                                style={[defaultStyles.leftViewContainer, styles.leftViewContainer]}
                             >
                                 {leftIcon &&
-                                    <TouchableWithoutFeedback onPress={this._onLeftIconPressed}>
+                                    <TouchableWithoutFeedback onPress={this.onLeftIconPressed}>
                                         <View style={[defaultStyles.leftIcon, styles.leftIcon]}>
                                             {leftIcon}
                                         </View>
                                     </TouchableWithoutFeedback>
                                 }
                                 {leftAvatar &&
-                                    <TouchableWithoutFeedback onPress={this._onLeftIconPressed}>
+                                    <TouchableWithoutFeedback onPress={this.onLeftIconPressed}>
                                         <View style={[defaultStyles.leftAvatar, styles.leftAvatar]}>
                                             {leftAvatar}
                                         </View>
                                     </TouchableWithoutFeedback>
                                 }
                             </View>
-                        }
+                    }
 
-                        <View style={defaultStyles.textViewContainer}>
-                            <View style={defaultStyles.firstLine}>
-                                <View style={defaultStyles.primaryTextContainer}>
-                                    <Text
-                                      numberOfLines={lines && lines === 'dynamic' ? null : 1}
-                                      style={[defaultStyles.primaryText, { color: primaryColor }]}
-                                    >
-                                        {primaryText}
-                                    </Text>
-                                </View>
-                            </View>
-                            {secondaryText &&
-                                <View>
-                                    <Text
-                                      style={[
-                                          { height: 18 },
-                                          lines > 2 && { height: 22 * (lines - 1) - 4 },
-                                          lines === 'dynamic' && { height: null },
-                                          defaultStyles.secondaryText
-                                      ]}
-                                    >
-                                        {secondaryText}
-                                    </Text>
-                                </View>
-                            }
-                            {secondaryTextMoreLine &&
-                                <View
-                                  style={[
-                                      { height: 18 },
-                                      lines > 2 && { height: 22 * (lines - 1) - 4 }
-                                  ]}
+                    <View style={defaultStyles.textViewContainer}>
+                        <View style={defaultStyles.firstLine}>
+                            <View style={defaultStyles.primaryTextContainer}>
+                                <Text
+                                    numberOfLines={lines && lines === 'dynamic' ? null : 1}
+                                    style={[defaultStyles.primaryText, { color: primaryColor }]}
                                 >
-                                    {secondaryTextMoreLine.map((line) => (
-                                        <Text style={[
+                                    {primaryText}
+                                </Text>
+                            </View>
+                        </View>
+                        {secondaryText &&
+                            <View>
+                                <Text
+                                    style={[
+                                      { height: 18 },
+                                        lines > 2 && { height: 22 * (lines - 1) - 4 },
+                                        lines === 'dynamic' && { height: null },
+                                        defaultStyles.secondaryText,
+                                    ]}
+                                >
+                                    {secondaryText}
+                                </Text>
+                            </View>
+                        }
+                        {secondaryTextMoreLine &&
+                            <View
+                                style={[
+                                  { height: 18 },
+                                    lines > 2 && { height: 22 * (lines - 1) - 4 },
+                                ]}
+                            >
+                                {secondaryTextMoreLine.map((line) => (
+                                    <Text
+                                        style={[
                                             defaultStyles.secondaryText,
                                             { height: 22 },
-                                            line.style]}
-                                        >
-                                            {line.text}
-                                        </Text>
-                                    ))}
-                                </View>
-                            }
-                        </View>
+                                            line.style,
+                                        ]}
+                                    >
+                                        {line.text}
+                                    </Text>
+                                ))}
+                            </View>
+                        }
+                    </View>
 
 
-                        <View style={[defaultStyles.rightViewContainer, styles.rightViewContainer]}>
-                            {rightAvatar &&
-                                <View style={[defaultStyles.rightAvatar]} >
-                                    {rightAvatar}
-                                </View>
-                            }
-                            {
-                                lines === 3 && !!rightIcon && !!captionText &&
-                                <View style={[defaultStyles.captionTextContainer]} >
+                    <View style={[defaultStyles.rightViewContainer, styles.rightViewContainer]}>
+                        {rightAvatar &&
+                            <View style={[defaultStyles.rightAvatar]} >
+                                {rightAvatar}
+                            </View>
+                        }
+                        {
+                            lines === 3 && !!rightIcon && !!captionText &&
+                                <View style={[defaultStyles.captionTextContainer]}>
                                     <Text style={[defaultStyles.captionText]}>{captionText}</Text>
                                 </View>
-                            }
-                            <View style={[defaultStyles.rightIcon, styles.rightIcon]} >
-                                {
-                                    menuActions &&
+                        }
+                        <View style={[defaultStyles.rightIcon, styles.rightIcon]} >
+                            {
+                                menuActions &&
                                     <IconToggle
-                                      color={primaryColor}
-                                      onPress={this._onMenuPressed}
+                                        color={primaryColor}
+                                        onPress={this.onMenuPressed}
                                     >
-                                        <Icon name={'more-vert'}
-                                          color={primaryColor}
-                                          ref="menu"
-                                          size={24}
+                                        <Icon
+                                            name="more-vert"
+                                            color={primaryColor}
+                                            ref="menu"
+                                            size={24}
                                         />
                                     </IconToggle>
-                                }
-                                {
-                                    rightIcon && !menuActions &&
+                            }
+                            {
+                                rightIcon && !menuActions &&
                                     <IconToggle
-                                      color={primaryColor}
-                                      onPress={onRightIconClick}
+                                        color={primaryColor}
+                                        onPress={onRightIconClick}
                                     >
                                         {rightIcon}
                                     </IconToggle>
-                                }
-                            </View>
+                            }
                         </View>
                     </View>
                 </View>
+            </View>
         );
 
         if (onPress) {
             return (
-                <TouchableNativeFeedback onPress={this._onListItemPressed}>
+                <TouchableNativeFeedback onPress={this.onListItemPressed}>
                     {content}
                 </TouchableNativeFeedback>
             );
