@@ -3,7 +3,7 @@ import {
     View,
     Text,
     TouchableWithoutFeedback,
-    TouchableNativeFeedback
+    TouchableNativeFeedback,
 } from 'react-native';
 import { TYPO, COLOR } from '../config';
 import Icon from '../Icon';
@@ -32,7 +32,7 @@ const propTypes = {
         PropTypes.shape({
             primaryText: PropTypes.string.isRequired,
             secondaryText: PropTypes.string,
-        })
+        }),
     ]),
 
     // right side
@@ -40,7 +40,7 @@ const propTypes = {
         PropTypes.element,
         PropTypes.string,
     ]),
-    onRightElementPress: PropTypes.func
+    onRightElementPress: PropTypes.func,
 };
 const defaultProps = {
     lines: 1,
@@ -78,10 +78,10 @@ const defaultStyles = StyleSheet.create({
         ...TYPO.paperFontSubhead,
     },
     firstLine: {
-        flexDirection: 'row'
+        flexDirection: 'row',
     },
     primaryTextContainer: {
-        flex: 1
+        flex: 1,
     },
     secondaryText: Object.assign({}, TYPO.paperFontBody1, {
         lineHeight: 22,
@@ -164,7 +164,7 @@ class ListItem extends Component {
         if (key === 'listItemContainer' || key === 'contentViewContainer') {
             result.contentViewContainer = {};
             result.listItemContainer = {
-                height: getListItemHeight({ dense, leftElement, numberOfLines })
+                height: getListItemHeight({ dense, leftElement, numberOfLines }),
             };
 
             if (numberOfLines === 3) {
@@ -240,7 +240,7 @@ class ListItem extends Component {
     renderCenterElement = () => {
         const {
             lines,
-            centerElement
+            centerElement,
         } = this.props;
 
         let content = null;
@@ -264,8 +264,8 @@ class ListItem extends Component {
                     <View style={this.getStyle('firstLine')}>
                         <View style={this.getStyle('primaryTextContainer')}>
                             <Text
-                              numberOfLines={lines && lines === 'dynamic' ? null : 1}
-                              style={this.getStyle('primaryText')}
+                                numberOfLines={lines && lines === 'dynamic' ? null : 1}
+                                style={this.getStyle('primaryText')}
                             >
                                 {primaryText}
                             </Text>
@@ -304,9 +304,9 @@ class ListItem extends Component {
         if (elements) {
             content = elements.map((action, i) => (
                 <IconToggle
-                  key={`${action}${i}`}
-                  color={COLOR.paperGrey500.color}
-                  onPress={() => this.onRightElementPressed({ action })}
+                    key={`${action}${i}`}
+                    color={COLOR.paperGrey500.color}
+                    onPress={() => this.onRightElementPressed({ action })}
                 >
                     <Icon name={action} size={24} style={this.getStyle('icon')} />
                 </IconToggle>
@@ -331,13 +331,13 @@ class ListItem extends Component {
         const { onPress } = this.props;
 
         const content = (
-                <View style={this.getStyle('listItemContainer')}>
-                    <View style={this.getStyle('contentViewContainer')}>
-                        {this.renderLeftElement()}
-                        {this.renderCenterElement()}
-                        {this.renderRightElement()}
-                    </View>
+            <View style={this.getStyle('listItemContainer')}>
+                <View style={this.getStyle('contentViewContainer')}>
+                    {this.renderLeftElement()}
+                    {this.renderCenterElement()}
+                    {this.renderRightElement()}
                 </View>
+            </View>
         );
 
         if (onPress) {
