@@ -7,19 +7,19 @@ const defaultStyles = StyleSheet.create({
         flex: 1,
         flexDirection: 'row',
         height: 56,
-        paddingHorizontal: 16
+        paddingHorizontal: 16,
     },
 
     contentViewContainer: {
         flex: 1,
         flexDirection: 'row',
-        alignItems: 'center'
+        alignItems: 'center',
     },
 
 
     leftViewContainer: {
         flexDirection: 'column',
-        width: 56
+        width: 56,
     },
     leftIcon: {
     },
@@ -28,43 +28,43 @@ const defaultStyles = StyleSheet.create({
     },
 
     textViewContainer: {
-        flex: 1
+        flex: 1,
     },
     primaryText: Object.assign({}, TYPO.paperFontSubhead, { lineHeight: 24 }),
     secondaryText: Object.assign({}, TYPO.paperFontBody1, {
         lineHeight: 22,
         fontSize: 14,
-        color: 'rgba(0,0,0,.54)'
+        color: 'rgba(0,0,0,.54)',
     }),
     firstLine: {
-        flexDirection: 'row'
+        flexDirection: 'row',
     },
     primaryTextContainer: {
-        flex: 1
+        flex: 1,
     },
 
     rightViewContainer: {
         flexDirection: 'column',
-        paddingLeft: 16
+        paddingLeft: 16,
     },
     rightActionsViewContainer: {
-        flex: 1
+        flex: 1,
     },
     rightIcon: {
         flex: 1,
     },
     rightAvatar: {
-        width: 56
+        width: 56,
     },
 
     captionText: Object.assign({}, TYPO.paperFontCaption),
     captionTextContainer: {
-        flex: 1
+        flex: 1,
     },
     captionTextContainer2: {
         alignSelf: 'flex-end',
-        alignItems: 'flex-end'
-    }
+        alignItems: 'flex-end',
+    },
 });
 
 export default class List extends Component {
@@ -78,29 +78,27 @@ export default class List extends Component {
 
     static defaultProps = {
         lines: 1,
-        primaryColor: 'rgba(0,0,0,.87)'
+        primaryColor: 'rgba(0,0,0,.87)',
     };
     render() {
         const {
             primaryText,
             leftElement,
             rightElement,
-            styles
+            styles,
         } = this.props;
 
         return (
-            <TouchableNativeFeedback onPress={this._onListItemPressed}>
+            <TouchableNativeFeedback onPress={this.onListItemPressed}>
                 <View style={[defaultStyles.listContainer, styles.listContainer]}>
                     <View style={[defaultStyles.contentViewContainer]}>
                         {
                             (leftElement) &&
-                            <View
-                              style={[defaultStyles.leftViewContainer]}
-                            >
-                                <View style={[defaultStyles.leftIcon]}>
-                                    {leftElement}
+                                <View style={[defaultStyles.leftViewContainer]}>
+                                    <View style={[defaultStyles.leftIcon]}>
+                                        {leftElement}
+                                    </View>
                                 </View>
-                            </View>
                         }
 
                         <View style={defaultStyles.textViewContainer}>
