@@ -5,33 +5,33 @@ import React, { Component, PropTypes } from 'react';
 export default class MenuController extends Component {
 
     static childContextTypes = {
-        menuController: PropTypes.object
+        menuController: PropTypes.object,
     };
     static propTypes = {
         children: PropTypes.node,
-        style: PropTypes.object
+        style: PropTypes.object,
     };
 
     constructor(props) {
         super(props);
 
-        this._menus = new Map();
+        this.menus = new Map();
 
         this.state = {
-            openMenu: null
+            openMenu: null,
         };
     }
     getChildContext() {
         return {
-            menuController: this
+            menuController: this,
         };
     }
 
     add(menu) {
-        const lastId = this._menus.size;
+        const lastId = this.menus.size;
         const menuId = lastId + 1;
 
-        this._menus.set(menuId, menu);
+        this.menus.set(menuId, menu);
 
         return menuId;
     }
@@ -41,24 +41,24 @@ export default class MenuController extends Component {
     open = (menu, target) => {
         this.setState({
             openMenu: menu,
-            target
+            target,
         });
     }
     toggle = (menu, target) => {
         if (this.state.openMenu) {
             this.setState({
                 openMenu: null,
-                target: null
+                target: null,
             });
         } else {
             this.setState({
                 openMenu: menu,
-                target
+                target,
             });
         }
     }
     remove = (menu) => {
-        this._menus.delete(menu.getId());
+        this.menus.delete(menu.getId());
     }
 
     render() {
@@ -72,7 +72,7 @@ export default class MenuController extends Component {
                 <View style={this.props.style}>
                     {this.props.children}
                 </View>
-                { overlay }
+                {overlay}
                 {
                     this.state.openMenu &&
                     this.state.openMenu.getOptions(this.state.target)
