@@ -8,11 +8,11 @@ const styles = StyleSheet.create({
         flexDirection: 'row',
         height: 48,
         paddingLeft: 16,
-        paddingRight: 16
+        paddingRight: 16,
     },
     itemText: {
-        fontSize: 16
-    }
+        fontSize: 16,
+    },
 });
 
 export default class MenuItem extends Component {
@@ -21,11 +21,11 @@ export default class MenuItem extends Component {
         closeMenu: PropTypes.func,
         label: PropTypes.string,
         onPress: PropTypes.func,
-        value: PropTypes.object
+        value: PropTypes.object,
     };
 
 
-    _onPressed = () => {
+    onPressed = () => {
         const { closeMenu, onPress, value } = this.props;
         closeMenu();
 
@@ -38,7 +38,7 @@ export default class MenuItem extends Component {
         const { label } = this.props;
 
         return (
-            <TouchableNativeFeedback onPress={this._onPressed}>
+            <TouchableNativeFeedback onPress={this.onPressed}>
                 <View style={styles.itemContainer} >
                     <Text style={styles.itemText}>{label}</Text>
                 </View>
