@@ -11,12 +11,12 @@ const styles = StyleSheet.create({
         elevation: 8,
         position: 'absolute',
         top: 0,
-        left: 0
+        left: 0,
     },
     content: {
         paddingTop: 8,
-        paddingBottom: 8
-    }
+        paddingBottom: 8,
+    },
 });
 
 export default class MenuItems extends Component {
@@ -24,7 +24,7 @@ export default class MenuItems extends Component {
     static propTypes = {
         closeMenu: PropTypes.func,
         target: PropTypes.object,
-        items: PropTypes.array
+        items: PropTypes.array,
     };
 
     constructor(props) {
@@ -32,8 +32,8 @@ export default class MenuItems extends Component {
         this.state = {
             style: {
                 top: 0,
-                left: -1000
-            }
+                left: -1000,
+            },
         };
     }
 
@@ -57,8 +57,8 @@ export default class MenuItems extends Component {
                     this.setState({
                         style: {
                             top: top || 0,
-                            left: left || 0
-                        }
+                            left: left || 0,
+                        },
                     });
                 });
             });
