@@ -9,20 +9,20 @@ const styles = StyleSheet.create({
         top: 0,
         left: 0,
         width: window.width,
-        height: window.height
+        height: window.height,
     },
     overlay: {
         position: 'absolute',
         backgroundColor: 'transparent',
         width: window.width,
-        height: window.height
-    }
+        height: window.height,
+    },
 });
 
 export default class Overlay extends Component {
 
     static propTypes = {
-        onPress: PropTypes.func
+        onPress: PropTypes.func,
     }
 
     render() {
