@@ -7,10 +7,10 @@ export default StyleSheet.create({
         elevation: 8,
         position: 'absolute',
         top: 0,
-        left: 0
+        left: 0,
     },
     content: {
         paddingTop: 8,
-        paddingBottom: 8
-    }
+        paddingBottom: 8,
+    },
 });
