@@ -13,19 +13,19 @@ export default class RadioButtonGroup extends Component {
         items: PropTypes.arrayOf(PropTypes.shape({
             value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
             label: PropTypes.string,
-            disabled: PropTypes.bool
-        }))
+            disabled: PropTypes.bool,
+        })),
     };
 
     static defaultProps = {
         theme: 'light',
-        primary: PRIMARY
+        primary: PRIMARY,
     };
 
     constructor(props) {
         super(props);
         this.state = {
-            selected: props.selected || null
+            selected: props.selected || null,
         };
     }
 
@@ -41,7 +41,7 @@ export default class RadioButtonGroup extends Component {
         const { onSelect } = this.props;
 
         this.setState({
-            selected: value
+            selected: value,
         });
 
         if (onSelect) {
@@ -74,14 +74,14 @@ export default class RadioButtonGroup extends Component {
                         const { value } = item;
                         return (
                             <RadioButton
-                              ref={value}
-                              key={`RadioButton${value}`}
-                              value={`${value.toString()}`}
-                              theme={theme}
-                              primary={primary}
-                              onSelect={this.onSelect}
-                              checked={this.state.selected && this.state.selected === value}
-                              {...item}
+                                ref={value}
+                                key={`RadioButton${value}`}
+                                value={`${value.toString()}`}
+                                theme={theme}
+                                primary={primary}
+                                onSelect={this.onSelect}
+                                checked={this.state.selected && this.state.selected === value}
+                                {...item}
                             />
                         );
                     })
