@@ -9,11 +9,11 @@ export default class Ripple extends Component {
         color: PropTypes.string,
         onPress: PropTypes.func,
         onLongPress: PropTypes.func,
-        children: PropTypes.node.isRequired
+        children: PropTypes.node.isRequired,
     };
 
     static defaultProps = {
-        color: 'rgba(0,0,0,.2)'
+        color: 'rgba(0,0,0,.2)',
     };
 
     render() {
@@ -25,10 +25,10 @@ export default class Ripple extends Component {
         if (!isCompatible('TouchableNativeFeedback')) {
             return (
                 <PolyfillRipple
-                  rippleColor={color}
-                  onPress={onPress}
-                  onLongPress={onLongPress}
-                  {...other}
+                    rippleColor={color}
+                    onPress={onPress}
+                    onLongPress={onLongPress}
+                    {...other}
                 >
                     {/* Stops fatal crash with out of bounds animation */}
                     <View style={{ marginHorizontal: .2 }}>
@@ -40,9 +40,9 @@ export default class Ripple extends Component {
 
         return (
             <TouchableNativeFeedback
-              background={TouchableNativeFeedback.Ripple(color)}
-              onPress={onPress}
-              onLongPress={onLongPress}
+                background={TouchableNativeFeedback.Ripple(color)}
+                onPress={onPress}
+                onLongPress={onLongPress}
             >
                 <View {...other}>
                     {children}
