@@ -5,9 +5,9 @@ import React, { Component, PropTypes } from 'react';
 
 const styles = StyleSheet.create({
     container: {
-        padding: 16
+        padding: 16,
     },
-    text: TYPO.paperFontBody1
+    text: TYPO.paperFontBody1,
 });
 
 export default class Subheader extends Component {
@@ -25,7 +25,7 @@ export default class Subheader extends Component {
         color: 'rgba(0,0,0,.54)',
         inset: false,
         theme: 'light',
-        lines: 1
+        lines: 1,
     };
 
     render() {
@@ -33,16 +33,16 @@ export default class Subheader extends Component {
 
         return (
             <View
-              style={[styles.container, {
-                  paddingLeft: inset ? 72 : 16
-              }, style]}
+                style={[styles.container, {
+                    paddingLeft: inset ? 72 : 16,
+                }, style]}
             >
                 <Text
-                  numberOfLines={lines}
-                  style={[styles.text, {
-                      color: getColor(color),
-                      fontWeight: '500'
-                  }]}
+                    numberOfLines={lines}
+                    style={[styles.text, {
+                        color: getColor(color),
+                        fontWeight: '500',
+                    }]}
                 >
                     {text}
                 </Text>
