@@ -20,18 +20,18 @@ const styles = StyleSheet.create({
     toolbar: {
         height: 56,
         flexDirection: 'row',
-        alignItems: 'center'
+        alignItems: 'center',
     },
     title: {
         flex: 1,
-        marginLeft: 16
+        marginLeft: 16,
     },
     leftIcon: {
-        margin: 16
+        margin: 16,
     },
     rightIcon: {
-        margin: 16
-    }
+        margin: 16,
+    },
 });
 
 export default class Toolbar extends Component {
@@ -49,7 +49,7 @@ export default class Toolbar extends Component {
             backgroundColor: PropTypes.string,
             titleColor: PropTypes.string,
             leftIconColor: PropTypes.string,
-            rightIconColor: PropTypes.string
+            rightIconColor: PropTypes.string,
         }),
         icon: PropTypes.string,
         onIconPress: PropTypes.func,
@@ -59,7 +59,7 @@ export default class Toolbar extends Component {
         })),
         menuActions: PropTypes.shape({
             onPress: PropTypes.func,
-            labels: PropTypes.array
+            labels: PropTypes.array,
         }),
         searchable: PropTypes.shape({
             onChangeText: PropTypes.func,
@@ -67,7 +67,7 @@ export default class Toolbar extends Component {
             placeholder: PropTypes.string,
             onSearchPressed: PropTypes.func,
             onSubmitEditing: PropTypes.func,
-            autoFocus: PropTypes.bool
+            autoFocus: PropTypes.bool,
         }),
         isSearchActive: PropTypes.bool,
     };
@@ -75,7 +75,7 @@ export default class Toolbar extends Component {
     static defaultProps = {
         theme: 'dark',
         primary: PRIMARY,
-        elevation: 4
+        elevation: 4,
     };
 
     constructor(props) {
@@ -83,22 +83,20 @@ export default class Toolbar extends Component {
 
         this.state = {
             isSearchActive: props.isSearchActive,
-            searchValue: ''
+            searchValue: '',
         };
     }
-
-    _handleShowPopupError() { }
-    _onMenuPressed = () => {
+    onMenuPressed = () => {
         const { menuActions } = this.props;
 
         UIManager.showPopupMenu(
             findNodeHandle(this.refs.menu),
             menuActions.labels,
-            this._handleShowPopupError,
+            () => {},
             menuActions.onPress
         );
     };
-    _onSearchTextChanged = (value) => {
+    onSearchTextChanged = (value) => {
         const { searchable } = this.props;
 
         if (isFunction(searchable.onChangeText)) {
@@ -107,7 +105,7 @@ export default class Toolbar extends Component {
 
         this.setState({ searchValue: value });
     };
-    _onSearchPressed = () => {
+    onSearchPressed = () => {
         const { searchable } = this.props;
 
         if (isFunction(searchable.onSearchPressed)) {
@@ -116,10 +114,10 @@ export default class Toolbar extends Component {
 
         this.setState({
             isSearchActive: true,
-            searchValue: ''
+            searchValue: '',
         });
     };
-    _onSearchClosePressed = () => {
+    onSearchClosePressed = () => {
         const { searchable } = this.props;
 
         if (isFunction(searchable.onSearchClosed)) {
@@ -128,13 +126,13 @@ export default class Toolbar extends Component {
 
         this.setState({
             isSearchActive: false,
-            searchValue: ''
+            searchValue: '',
         });
     };
 
 
     focusSearchField() {
-        this._searchFieldRef.focus();
+        this.searchFieldRef.focus();
     }
 
     render() {
@@ -152,7 +150,7 @@ export default class Toolbar extends Component {
             onIconPress,
             actions,
             menuActions,
-            searchable
+            searchable,
         } = this.props;
 
         const themeMap = {
@@ -160,19 +158,19 @@ export default class Toolbar extends Component {
                 backgroundColor: '#ffffff',
                 color: 'rgba(0,0,0,.87)',
                 leftIconColor: 'rgba(0,0,0,.54)',
-                rightIconColor: 'rgba(0,0,0,.54)'
+                rightIconColor: 'rgba(0,0,0,.54)',
             },
             dark: {
                 backgroundColor: getColor(primary),
                 color: 'rgba(255,255,255,.87)',
                 leftIconColor: 'rgba(255,255,255,.87)',
-                rightIconColor: 'rgba(255,255,255,.87)'
-            }
+                rightIconColor: 'rgba(255,255,255,.87)',
+            },
         };
 
         const opacityMap = {
             light: .38,
-            dark: .30
+            dark: .30,
         };
 
         // TODO: refactor this
@@ -188,7 +186,7 @@ export default class Toolbar extends Component {
                                 themeMap[theme].leftIconColor,
             rightIconColor: overrides && overrides.rightIconColor ?
                                 getColor(overrides.rightIconColor) :
-                                themeMap[theme].rightIconColor
+                                themeMap[theme].rightIconColor,
         };
 
         let iconMap = icon;
@@ -198,122 +196,126 @@ export default class Toolbar extends Component {
 
         if (this.state.isSearchActive) {
             iconMap = 'arrow-back';
-            onIconPressMap = this._onSearchClosePressed;
+            onIconPressMap = this.onSearchClosePressed;
             styleMap.backgroundColor = getColor('paperGrey100');
             styleMap.color = themeMap.light.color;
             styleMap.leftIconColor = themeMap.light.leftIconColor;
             styleMap.rightIconColor = themeMap.light.rightIconColor;
             actionsMap = [{
                 icon: 'clear',
-                onPress: () => this._onSearchTextChanged('')
+                onPress: () => this.onSearchTextChanged(''),
             }];
             menuActionsMap = null;
         }
 
         return (
             <View
-              style={[
-                  styles.toolbar,
+                style={[
+                    styles.toolbar,
                   { backgroundColor: styleMap.backgroundColor, elevation },
-                  style
-              ]}
+                    style,
+                ]}
             >
                 {
                     iconMap && (
                         <IconToggle
-                          color={styleMap.leftIconColor}
-                          onPress={onIconPressMap}
+                            color={styleMap.leftIconColor}
+                            onPress={onIconPressMap}
                         >
-                            <Icon name={iconMap || 'menu'}
-                              size={24}
-                              color={styleMap.leftIconColor}
-                              style={[styles.leftIcon, leftIconStyle]}
+                            <Icon
+                                name={iconMap || 'menu'}
+                                size={24}
+                                color={styleMap.leftIconColor}
+                                style={[styles.leftIcon, leftIconStyle]}
                             />
                         </IconToggle>
                     )
                 }
                 {
                     !this.state.isSearchActive &&
-                    <TouchableWithoutFeedback onPress={onTitlePress}>
-                        <View style={{ flex: 1, flexDirection: 'row' }}>
-                            <Text
-                              numberOfLines={1}
-                              style={[styles.title, TYPO.paperFontTitle, {
-                                  color: styleMap.color,
-                                  marginLeft: iconMap ? styles.title.marginLeft : 16
-                              }]}
-                            >
-                                {title}
-                            </Text>
-                        </View>
-                    </TouchableWithoutFeedback>
+                        <TouchableWithoutFeedback onPress={onTitlePress}>
+                            <View style={{ flex: 1, flexDirection: 'row' }}>
+                                <Text
+                                    numberOfLines={1}
+                                    style={[styles.title, TYPO.paperFontTitle, {
+                                        color: styleMap.color,
+                                        marginLeft: iconMap ? styles.title.marginLeft : 16,
+                                    }]}
+                                >
+                                    {title}
+                                </Text>
+                            </View>
+                        </TouchableWithoutFeedback>
                 }
                 {
                     this.state.isSearchActive &&
-                    <TextInput
-                      ref={(ref) => { this._searchFieldRef = ref; }}
-                      autoFocus={searchable.autoFocus}
-                      onChangeText={this._onSearchTextChanged}
-                      onSubmitEditing={searchable.onSubmitEditing}
-                      placeholder={searchable.placeholder}
-                      style={[styles.title, TYPO.paperFontTitle, {
-                          fontWeight: 'normal',
-                          color: styleMap.color,
-                          marginLeft: iconMap ? styles.title.marginLeft : 16
-                      }]}
-                      underlineColorAndroid="transparent"
-                      value={this.state.searchValue}
-                    />
+                        <TextInput
+                            ref={(ref) => { this.searchFieldRef = ref; }}
+                            autoFocus={searchable.autoFocus}
+                            onChangeText={this.onSearchTextChanged}
+                            onSubmitEditing={searchable.onSubmitEditing}
+                            placeholder={searchable.placeholder}
+                            style={[styles.title, TYPO.paperFontTitle, {
+                                fontWeight: 'normal',
+                                color: styleMap.color,
+                                marginLeft: iconMap ? styles.title.marginLeft : 16,
+                            }]}
+                            underlineColorAndroid="transparent"
+                            value={this.state.searchValue}
+                        />
                 }
                 {
                     actionsMap &&
                     actionsMap.map((action, i) => (
-                            <IconToggle
-                              key={i}
-                              color={styleMap.rightIconColor}
-                              badge={action.badge}
-                              onPress={action.onPress}
-                              disabled={action.disabled}
-                            >
-                                <Icon name={action.icon}
-                                  size={24}
-                                  color={styleMap.rightIconColor}
-                                  style={[
-                                      styles.rightIcon,
-                                      rightIconStyle,
-                                      action.disabled ? { opacity: opacityMap[theme] } : null
-                                  ]}
-                                />
-                            </IconToggle>
+                        <IconToggle
+                            key={i}
+                            color={styleMap.rightIconColor}
+                            badge={action.badge}
+                            onPress={action.onPress}
+                            disabled={action.disabled}
+                        >
+                            <Icon
+                                name={action.icon}
+                                size={24}
+                                color={styleMap.rightIconColor}
+                                style={[
+                                    styles.rightIcon,
+                                    rightIconStyle,
+                                    action.disabled ? { opacity: opacityMap[theme] } : null,
+                                ]}
+                            />
+                        </IconToggle>
                         )
                     )
                 }
                 {
                     searchable && !this.state.isSearchActive &&
-                    <IconToggle
-                      color={styleMap.rightIconColor}
-                      onPress={this._onSearchPressed}
-                    >
-                        <Icon name={'search'}
-                          color={styleMap.rightIconColor}
-                          size={24}
-                          style={[styles.rightIcon, rightIconStyle]}
-                        />
-                    </IconToggle>
+                        <IconToggle
+                            color={styleMap.rightIconColor}
+                            onPress={this.onSearchPressed}
+                        >
+                            <Icon
+                                name={'search'}
+                                color={styleMap.rightIconColor}
+                                size={24}
+                                style={[styles.rightIcon, rightIconStyle]}
+                            />
+                        </IconToggle>
                 }
                 {
                     menuActionsMap &&
-                    <IconToggle
-                      color={styleMap.rightIconColor}
-                      onPress={this._onMenuPressed}
-                    >
-                        <Icon name={'more-vert'}
-                          color={styleMap.rightIconColor}
-                          ref="menu"
-                          size={24}
-                          style={[styles.rightIcon, rightIconStyle]}
-                        />
-                    </IconToggle>
+                        <IconToggle
+                            color={styleMap.rightIconColor}
+                            onPress={this.onMenuPressed}
+                        >
+                            <Icon
+                                name={'more-vert'}
+                                color={styleMap.rightIconColor}
+                                ref="menu"
+                                size={24}
+                                style={[styles.rightIcon, rightIconStyle]}
+                            />
+                        </IconToggle>
                 }
             </View>
         );
