@@ -32,7 +32,7 @@ const propTypes = {
         placeholder: PropTypes.string,
         onSearchPressed: PropTypes.func,
         onSubmitEditing: PropTypes.func,
-        autoFocus: PropTypes.bool
+        autoFocus: PropTypes.bool,
     }),
     style: PropTypes.shape({
         toolbarContainer: PropTypes.object,
@@ -111,7 +111,7 @@ const styles = StyleSheet.create({
     },
     rightIcon: {
         margin: 16,
-    }
+    },
 });
 
 class Toolbar extends Component {
@@ -120,7 +120,7 @@ class Toolbar extends Component {
 
         this.state = {
             isSearchActive: props.isSearchActive,
-            searchValue: ''
+            searchValue: '',
         };
     }
     onMenuPressed = (labels) => {
@@ -151,7 +151,7 @@ class Toolbar extends Component {
 
         this.setState({
             isSearchActive: true,
-            searchValue: ''
+            searchValue: '',
         });
     };
     onSearchClosePressed = () => {
@@ -163,7 +163,7 @@ class Toolbar extends Component {
 
         this.setState({
             isSearchActive: false,
-            searchValue: ''
+            searchValue: '',
         });
     };
     getStyle = () => {
@@ -176,14 +176,14 @@ class Toolbar extends Component {
                 backgroundColor: '#ffffff',
                 color: 'rgba(0,0,0,.87)',
                 leftIconColor: 'rgba(0,0,0,.54)',
-                rightIconColor: 'rgba(0,0,0,.54)'
+                rightIconColor: 'rgba(0,0,0,.54)',
             };
         } else {
             selectedTheme = {
                 backgroundColor: getColor(primary),
                 color: 'rgba(255,255,255,.87)',
                 leftIconColor: 'rgba(255,255,255,.87)',
-                rightIconColor: 'rgba(255,255,255,.87)'
+                rightIconColor: 'rgba(255,255,255,.87)',
             };
         }
 
@@ -219,34 +219,34 @@ class Toolbar extends Component {
         result.centerElementContainer = [
             styles.centerElementContainer,
             searchStyle.centerElementContainer,
-            style.centerElementContainer
+            style.centerElementContainer,
         ];
         result.titleText = [
             styles.titleText,
             TYPO.paperFontTitle,
             { color: selectedTheme.color },
             searchStyle.titleText,
-            style.titleText
+            style.titleText,
         ];
         result.leftElementContainer = [
             styles.leftElementContainer,
             searchStyle.leftElementContainer,
-            style.leftElementContainer
+            style.leftElementContainer,
         ];
         result.rightElementContainer = [
             styles.rightElementContainer,
             searchStyle.rightElementContainer,
-            style.rightElementContainer
+            style.rightElementContainer,
         ];
         result.leftIcon = [
             styles.leftIcon,
             searchStyle.leftIcon,
-            style.leftIcon
+            style.leftIcon,
         ];
         result.rightIcon = [
             styles.rightIcon,
             searchStyle.rightIcon,
-            style.rightIcon
+            style.rightIcon,
         ];
         result.iconProps = {
             size: iconProps.size,
@@ -256,7 +256,7 @@ class Toolbar extends Component {
         return result;
     }
     focusSearchField() {
-        this._searchFieldRef.focus();
+        this.searchFieldRef.focus();
     }
     renderLeftElement = (style) => {
         const { leftElement, onLeftElementPress } = this.props;
@@ -277,10 +277,10 @@ class Toolbar extends Component {
             <View style={style.leftElementContainer}>
                 <IconToggle color={style.iconProps.color} onPress={onPress}>
                     <Icon
-                      name={iconName}
-                      size={style.iconProps.size}
-                      color={style.iconProps.color}
-                      style={style.leftIcon}
+                        name={iconName}
+                        size={style.iconProps.size}
+                        color={style.iconProps.color}
+                        style={style.leftIcon}
                     />
                 </IconToggle>
             </View>
@@ -320,14 +320,14 @@ class Toolbar extends Component {
         return (
             <View style={style.centerElementContainer}>
                 <TextInput
-                  ref={(ref) => { this._searchFieldRef = ref; }}
-                  autoFocus={searchable.autoFocus}
-                  onChangeText={this.onSearchTextChanged}
-                  onSubmitEditing={searchable.onSubmitEditing}
-                  placeholder={searchable.placeholder}
-                  style={style.titleText}
-                  underlineColorAndroid="transparent"
-                  value={this.state.searchValue}
+                    ref={(ref) => { this.searchFieldRef = ref; }}
+                    autoFocus={searchable.autoFocus}
+                    onChangeText={this.onSearchTextChanged}
+                    onSubmitEditing={searchable.onSubmitEditing}
+                    placeholder={searchable.placeholder}
+                    style={style.titleText}
+                    underlineColorAndroid="transparent"
+                    value={this.state.searchValue}
                 />
             </View>
         );
@@ -352,7 +352,7 @@ class Toolbar extends Component {
         if (this.state.isSearchActive) {
             actionsMap = [{
                 icon: 'clear',
-                onPress: () => this.onSearchTextChanged('')
+                onPress: () => this.onSearchTextChanged(''),
             }];
         }
 
@@ -371,19 +371,19 @@ class Toolbar extends Component {
                 } else {
                     content = (
                         <Icon
-                          name={action}
-                          size={style.iconProps.size}
-                          color={style.iconProps.color}
-                          style={style.rightIcon}
+                            name={action}
+                            size={style.iconProps.size}
+                            color={style.iconProps.color}
+                            style={style.rightIcon}
                         />
                     );
                 }
 
                 return (
                     <IconToggle
-                      key={index}
-                      color={style.iconProps.color}
-                      onPress={() => onRightElementPress({ action, index })}
+                        key={index}
+                        color={style.iconProps.color}
+                        onPress={() => onRightElementPress({ action, index })}
                     >
                         {content}
                     </IconToggle>
@@ -394,14 +394,15 @@ class Toolbar extends Component {
         if (searchable && !this.state.isSearchActive) {
             result.push(
                 <IconToggle
-                  key="searchIcon"
-                  color={style.iconProps.color}
-                  onPress={this.onSearchPressed}
+                    key="searchIcon"
+                    color={style.iconProps.color}
+                    onPress={this.onSearchPressed}
                 >
-                    <Icon name="search"
-                      size={style.iconProps.size}
-                      color={style.iconProps.color}
-                      style={style.rightIcon}
+                    <Icon
+                        name="search"
+                        size={style.iconProps.size}
+                        color={style.iconProps.color}
+                        style={style.rightIcon}
                     />
                 </IconToggle>
             );
@@ -409,16 +410,16 @@ class Toolbar extends Component {
         if (rightElement.menu && !this.state.isSearchActive) {
             result.push(
                 <IconToggle
-                  key="menuIcon"
-                  color={style.iconProps.color}
-                  onPress={() => this.onMenuPressed(rightElement.menu.labels)}
+                    key="menuIcon"
+                    color={style.iconProps.color}
+                    onPress={() => this.onMenuPressed(rightElement.menu.labels)}
                 >
                     <Icon
-                      ref="menu"
-                      name="more-vert"
-                      size={style.iconProps.size}
-                      color={style.iconProps.color}
-                      style={style.rightIcon}
+                        ref="menu"
+                        name="more-vert"
+                        size={style.iconProps.size}
+                        color={style.iconProps.color}
+                        style={style.rightIcon}
                     />
                 </IconToggle>
             );
@@ -435,9 +436,9 @@ class Toolbar extends Component {
 
         return (
             <Animated.View style={style.toolbarContainer}>
-                { this.renderLeftElement(style) }
-                { this.renderCenterElement(style) }
-                { this.renderRightElement(style) }
+                {this.renderLeftElement(style)}
+                {this.renderCenterElement(style)}
+                {this.renderRightElement(style)}
             </Animated.View>
         );
     }
