@@ -26,7 +26,7 @@ export const PRIMARY_COLORS = [
     'googleBlue', 'googleGreen', 'googleGrey', 'googleRed', 'googleYellow', 'paperAmber',
     'paperBlue', 'paperBlueGrey', 'paperBrown', 'paperCyan', 'paperDeepOrange', 'paperDeepPurple',
     'paperGreen', 'paperGrey', 'paperIndigo', 'paperLightBlue', 'paperLightGreen', 'paperLime',
-    'paperOrange', 'paperPink', 'paperPurple', 'paperRed', 'paperTeal', 'paperYellow'
+    'paperOrange', 'paperPink', 'paperPurple', 'paperRed', 'paperTeal', 'paperYellow',
 ];
 
 /**
