@@ -4,7 +4,7 @@ import React, { Component, PropTypes } from 'react';
 
 const styles = {
     container: {
-        overflow: 'hidden'
+        overflow: 'hidden',
     },
     background: {
         flex: 1,
@@ -12,11 +12,11 @@ const styles = {
         top: 0,
         right: 0,
         bottom: 0,
-        left: 0
+        left: 0,
     },
     ripple: {
-        position: 'absolute'
-    }
+        position: 'absolute',
+    },
 };
 
 export default class Ripple extends Component {
@@ -27,12 +27,12 @@ export default class Ripple extends Component {
         onPress: PropTypes.func,
         onLongPress: PropTypes.func,
         style: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
-        children: PropTypes.element.isRequired
+        children: PropTypes.element.isRequired,
     };
 
     static defaultProps = {
         rippleColor: 'rgba(0,0,0,.2)',
-        elevation: null
+        elevation: null,
     };
 
     constructor(props) {
@@ -46,16 +46,25 @@ export default class Ripple extends Component {
             size: null,
             location: null,
             longPress: false,
-            elevation: props.elevation ? props.elevation[0] : null
+            elevation: props.elevation ? props.elevation[0] : null,
         };
     }
 
-    _highlight = (e) => {
+    getContainerDimensions = (next) => {
+        this.refs.container.measure((x, y, width, height, pageX, pageY) => {
+            this.setState({
+                size: 3 * (width > height ? width : height),
+                location: { pageX, pageY },
+            }, next);
+        });
+    };
+
+    highlight = (e) => {
         const { elevation } = this.props;
 
         if (elevation) {
             this.setState({
-                elevation: elevation[1]
+                elevation: elevation[1],
             });
         }
 
@@ -64,40 +73,40 @@ export default class Ripple extends Component {
         this.setState({
             rippling: true,
             pageX,
-            pageY
+            pageY,
         });
 
-        this._getContainerDimensions(() => {
+        this.getContainerDimensions(() => {
             const duration = (this.state.size / 100) * 110;
 
             Animated.timing(this.state.scaleValue, {
                 toValue: 1,
-                duration: duration < 500 || duration >= 1500 ? 500 : duration
+                duration: duration < 500 || duration >= 1500 ? 500 : duration,
             }).start();
 
             Animated.timing(this.state.fadeValue, {
                 toValue: .2,
-                duration: 200
+                duration: 200,
             }).start();
         });
     };
 
-    _unHighlight = () => {
+    unHighlight = () => {
         const { elevation } = this.props;
 
         if (elevation) {
             this.setState({
-                elevation: elevation[0]
+                elevation: elevation[0],
             });
         }
 
         this.setState({
-            rippling: false
+            rippling: false,
         });
 
         Animated.timing(this.state.scaleValue, {
             toValue: 0.001,
-            duration: 100
+            duration: 100,
         }).start();
 
         Animated.timing(this.state.fadeValue, {
@@ -105,22 +114,13 @@ export default class Ripple extends Component {
             duration: 100,
         }).start();
     };
-
-    _getContainerDimensions = (next) => {
-        this.refs.container.measure((x, y, width, height, pageX, pageY) => {
-            this.setState({
-                size: 3 * (width > height ? width : height),
-                location: { pageX, pageY }
-            }, next);
-        });
-    };
     render() {
         const {
             rippleColor,
             onPress,
             onLongPress,
             children,
-            style
+            style,
         } = this.props;
         const {
             fadeValue,
@@ -130,43 +130,43 @@ export default class Ripple extends Component {
             rippling,
             scaleValue,
             location,
-            elevation
+            elevation,
         } = this.state;
 
         return (
             <TouchableOpacity
-              ref="container"
-              activeOpacity={1}
-              onPress={onPress}
-              onLongPress={onLongPress}
-              onPressIn={this._highlight}
-              onPressOut={this._unHighlight}
+                ref="container"
+                activeOpacity={1}
+                onPress={onPress}
+                onLongPress={onLongPress}
+                onPressIn={this.highlight}
+                onPressOut={this.unHighlight}
             >
                 <View
-                  style={style || {}}
-                  elevation={elevation || 0}
+                    style={style || {}}
+                    elevation={elevation || 0}
                 >
-                    <Animated.View style={[
-                        styles.background, {
-                            backgroundColor: rippling ? rippleColor : 'transparent',
-                            opacity: fadeValue
-                        }
-                    ]}
+                    <Animated.View
+                        style={[
+                            styles.background, {
+                                backgroundColor: rippling ? rippleColor : 'transparent',
+                                opacity: fadeValue,
+                            },
+                        ]}
                     />
-                    <Animated.View style={[
-                        styles.ripple,
-                        location &&
-                            {
+                    <Animated.View
+                        style={[
+                            styles.ripple,
+                            location && {
                                 backgroundColor: rippleColor,
                                 width: size,
                                 height: size,
                                 top: pageY - location.pageY - size / 2,
                                 left: pageX - location.pageX - size / 2,
-                                borderRadius: size / 2
+                                borderRadius: size / 2,
                             },
-                        {
-                            transform: [{ scale: scaleValue }]
-                        }]}
+                            { transform: [{ scale: scaleValue }] },
+                        ]}
                     />
                     {children}
                 </View>
